changes to main.py and services

+++ server/app/services.py
@@ -49,12 +49,14 @@ def delete_user(user_id):
     user = User.get(id=user_id)
     #if user is found get community associated with user in user.community
     if user:
-        community = Community.get(id=user.community.id)
-        if community:
-            jug_user = JugUser.get(id=user.community.id)
-            if jug_user:
-                jug_user.delete()
-            community.delete()
+        community_id = user.community
+        if community_id:
+            community = Community.get(id=user.community.id)
+            if community:
+                jug_user = JugUser.get(id=user.community.id)
+                if jug_user:
+                    jug_user.delete()
+                community.delete()
         user.delete()
